requirements.txt
Django==2.1.3
django-localflavor==2.1
djangorestframework==3.9.0
Faker==1.0.0
model-mommy==1.6.0
Pillow==5.3.0
python-dateutil==2.7.5
pytz==2018.7
six==1.11.0
text-unidecode==1.2
